// src/app/core/models/house-response.model.ts
export interface HouseResponse {
  id: number;
  name: string;
  description: string;
  category: string;
  numberOfRooms: number;
  numberOfBathrooms: number;
  price: number;
  address: string;
  ubication: string;
  city: string;
  department: string;
  activePublicationDate: string;
  publicationStatus: string;
  publicationDate: string;
  sellerId: number;
}